Feature/3.0 liaohj

+++ include/common/taosmsg.h
@@ -281,7 +281,7 @@ typedef struct SSchema {
   uint8_t type;
   char    name[TSDB_COL_NAME_LEN];
   int16_t colId;
-  int16_t bytes;
+  int32_t bytes;
 } SSchema;
 //#endif
